refactor: Admin input date styles moved to component scope
