+++ src/Admin/AdminForm/AdminFormInputs/index.tsx
@@ -3,7 +3,7 @@ import AdminInputCelsius from './AdminInputCelsius'
 import AdminInputChips from './AdminInputChips/AdminInputChips'
 import AdminInputCode from './AdminInputCode'
 import AdminInputCountry from './AdminInputCountry'
-import AdminInputDate from './AdminInputDate'
+import AdminInputDate from './AdminInputDate/AdminInputDate'
 import AdminInputEmail from './AdminInputEmail'
 import AdminInputJson from './AdminInputJson'
 import AdminInputLink from './AdminInputLink'
